Split out constants from API library
